Use a buffer of recent updates to determine ETA
When eta is calculated only based on the most recent update then it fluctuates wildly to the point of being totally unusable - especially when iterating over a large set, and sending an update event on every iteration.

Instead save a buffer of the values and timestamps of all updates and calculate the rate of change - and hence eta - based on the moving average of the last N updates based on the option passed.