Fix frames-per-second parsing
The calculation of the throttle time based on `fps` option was the wrong way round so that *decreasing* the fps (which *should* result in fewer renders) actually caused the render rate to increase.

A `fps` of 10 should result in a timeout value of 100ms, but as the code was previously it resulted in a timeout of 0.01ms, which effectively meant that there was no throttling.